optimization/requirements.txt
firebase-admin
numpy
python-dotenv
googlemaps
